+++ src/ska_dishln_pointing_device/commands/generate_program_track_table.py
@@ -33,6 +33,7 @@ class GenerateProgramTrackTable(FastCommand):
     def do(self, *args, **kwargs) -> None:
         """This method generates program track table."""
         try:
+            self.logger.debug("Executing GenerateProgramTrackTable command.")
             with self.component_manager.track_thread_lock:
                 self.component_manager.mapping_scan_event.clear()
             if (