+++ src/ska_dishln_pointing_device/commands/stop_program_track_table.py
@@ -31,8 +31,15 @@ class StopProgramTrackTable(FastCommand):
     def do(self, *args, **kwargs) -> None:
         """This method stops program track table generation."""
         try:
+            self.logger.debug("Executing StopProgramTrackTable command.")
             with self.component_manager.track_thread_lock:
                 self.component_manager.mapping_scan_event.set()
+            if (
+                self.component_manager.track_table_thread
+                and self.component_manager.track_table_thread.is_alive()
+            ):
+                self.component_manager.track_table_thread.join()
+            self.component_manager.update_program_track_table([])
             self.logger.info(
                 "StopProgramTrackTable command executed successfully"
             )