+++ src/ska_tmc_dishleafnode/commands/configure_band_command.py
@@ -100,6 +100,9 @@ class ConfigureBand(DishLNCommand):
         return:
             (ResultCode, str)
         """
+        self.logger.debug(
+            "Input argument for ConfigureBand command is: %s", argin
+        )
         result_code, message = self.init_adapter()
         if result_code == ResultCode.FAILED:
             self.logger.error(